[#671] Almost done checking contents
